refactor(design-system): promote SliderInput into design-system
SliderInput (label + slider + editable value badge) is a generic composite
built entirely on design-system primitives (Slider, cn, variants) with all
text supplied via props — a clean fit for the design system. Move it from
`shared/components` into `design-system` and export it from the public index.

- Migrated all 8 call sites to import from `@/design-system`.
- Removed the bin-designer `controls/SliderInput` re-export shim and its
  barrel export.
- Repointed the four test files that stubbed SliderInput to partially mock
  `@/design-system` (spread-actual + override).
- Folded the two unique cases from the deleted shim test (clamp-to-min,
  NaN input) into the canonical design-system test so coverage is preserved.

SnappingSlider and LidExplodeSlider remain feature-local for now — both are
bin-designer-only, already build on the design-system SliderThumb primitive,
and carry domain-coupled behaviour (i18n strings, hardcoded tick waypoints)
that warrants a deliberate API design in a follow-up rather than a rushed lift.
