+++ src/design-system/SliderInput/index.ts
