--- src/features/bin-designer/components/controls/SliderInput/SliderInput.tsx
@@ -1,7 +1,0 @@
-/**
- * Re-export shared SliderInput for bin-designer backward compatibility.
- *
- * The canonical implementation lives in @/shared/components/SliderInput.
- */
-
-export { SliderInput } from '@/shared/components/SliderInput';
