--- src/features/bin-designer/components/controls/SliderInput/index.ts
@@ -1,1 +1,0 @@
-export * from './SliderInput';
