+++ src/features/bin-designer/components/controls/index.ts
@@ -1,4 +1,3 @@
-export { SliderInput } from './SliderInput';
 export { SnappingSlider } from './SnappingSlider';
 export type { SnappingSliderOption } from './SnappingSlider';
 export { ThicknessSelector } from './ThicknessSelector';
