+++ src/features/bin-designer/components/panel/CutoutsSection/CutoutEditor.test.tsx
@@ -1,3 +1,4 @@
+import type * as DesignSystem from '@/design-system';
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { render, screen } from '@testing-library/react';
 import { useDesignerStore } from '@/features/bin-designer/store';
@@ -8,7 +9,8 @@ vi.mock('@/i18n', () => ({
   useTranslation: () => (key: string) => key,
 }));
 
-vi.mock('../../controls/SliderInput', () => ({
+vi.mock('@/design-system', async () => ({
+  ...(await vi.importActual<typeof DesignSystem>('@/design-system')),
   SliderInput: ({ label }: { label: string }) => <div data-testid={`slider-${label}`} />,
 }));
