+++ src/features/bin-designer/components/panel/OverhangSection/OverhangSection.tsx
@@ -8,8 +8,7 @@
  * bins.
  */
 
-import { SliderInput } from '@/shared/components/SliderInput';
-import { Checkbox } from '@/design-system';
+import { Checkbox, SliderInput } from '@/design-system';
 import { CollapsibleSection } from '@/shared/components/CollapsibleSection';
 import { DESIGNER_CONSTRAINTS } from '../../../constants';
 import { FeatureGate } from '../FeatureGate';
